add support for `table`s
recursively generate members using the json representation as the id for the reference.

there are some interesting side effects to this approach. since everything is effectively serialized to its value you can't tell if two objects are actually the same instance or just share the same value. this reduces duplication, but may actually be misleading to the frontend...